refactor(frontend): group recipes MFE cross-backend code in integrations/
The recipes MFE owns the primary "recipes" surface but also drives
cross-backend flows: assign-to-meal-plan, create-shopping-list-from-
recipes, chat-based recipe search. Those services and dialogs were
scattered across recipe-list/ and recipe-detail/, mixed in with files
that belonged to the pure recipes feature itself.

Pulled them into a new integrations/ subtree so the cross-MFE coupling
is visible at the folder level:

  recipes/src/app/
  ├── (pure recipe features stay): cook-mode/, cookable-recipes/,
  │   recipe-detail/, recipe-edit/, recipe-list/
  └── integrations/
      ├── shopping/
      │   ├── create-shopping-list-dialog/      ← from recipe-detail/
      │   ├── multi-recipe-preview-dialog/      ← from recipe-list/
      │   └── multi-recipe-shopping-list.service*
      ├── meal-plan/
      │   └── recipe-assignment.service*        ← from recipe-list/
      └── chat/
          └── recipe-chat-search.service*       ← from recipe-list/

Same git mv pattern as the api-lib split — history follows. Imports
in recipe-list.component / recipe-detail.component re-routed to
../integrations/*. Moved files needed their `../api` paths bumped to
`../../api` because they're now one folder deeper.

Also added `.nx/coverage` to client/.prettierignore (Vitest writes
coverage HTML there when run locally; gitignored so CI never saw it,
but blocked the pre-commit prettier check).

recipes MFE tests + lint + build all pass.
