+++ client/apps/recipes/src/app/integrations/shopping/create-shopping-list-dialog/create-shopping-list-dialog.component.html
