+++ client/apps/recipes/src/app/integrations/shopping/multi-recipe-preview-dialog/multi-recipe-preview-dialog.component.scss
